Created rust websocket server that works with app frontend. This breaks the docker for know I will fix it later

+++ backend/migrations/0001_create-database.sql
@@ -1,6 +1,6 @@
 CREATE TABLE canvas (
-  x smallint,
-  y smallint,
-  colour smallint NOT NULL,
+  x integer,
+  y integer,
+  colour integer NOT NULL,
   PRIMARY KEY(x, y)
 );
